requirements.txt
pytorchvideo==0.1.5
duckdb==0.6.1
torch-tb-profiler==0.4.0
